Numbers: * Move PrimeCacheArgumentProvider.java to primes.cache package + update usages
