+++ numbers/src/test/java/mathtools/numbers/primes/cache/PrimeCacheArgumentProvider.java
@@ -1,10 +1,13 @@
-package mathtools.numbers.primes;
+package mathtools.numbers.primes.cache;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 
 import java.util.stream.Stream;
+
+import mathtools.numbers.primes.BytePrimeCache;
+import mathtools.numbers.primes.ShortPrimeCache;
 
 /** ArgumentProvider for all [PrimeCacheBase] implementations
  * @author DK96-OS : 2022 */
